ChatWithPrivateDoc_2.0/requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchain-community
langchainhub
pinecone-client
python-dotenv
tiktoken
wikipedia
pypdf
docx2txt
chromadb
langchain_openai
streamlit
requests
pandas
numpy
